+++ src/main/java/com/devland/walletapi/transaction/TransactionRepository.java
@@ -10,10 +10,10 @@ import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface TransactionRepository extends JpaRepository<Transaction, BigInteger> {
+public interface TransactionRepository extends JpaRepository<Transaction, Long> {
     // JPQL
     @Query("SELECT t FROM Transaction t WHERE t.amount = ?1 and t.id = ?2")
-    public List<Transaction> ngambilTransaksi(double amount, BigInteger id);
+    public List<Transaction> ngambilTransaksi(double amount, Long id);
 
     // = ? -> query binding //
     //    @Query(value = "SELECT * FROM transaction WHERE amount = ?1 and id = ?2"
